nonebot_plugin_easy_translate/pyproject.toml
[tool.poetry]
name = "nonebot_plugin_easy_translate"
version = "0.1.0"
description = "Nonebot2 免api key简单翻译插件"
authors = ["nikissXI <1299577815@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_easy_translate"}]
homepage = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_easy_translate"
repository = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_easy_translate"
documentation = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_easy_translate#README.md"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0-rc.1"
nonebot-adapter-onebot = "^2.0.0"
httpx = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
